important values types changed from double to BigDecimal

+++ src/main/java/currency/ExchangeData.java
@@ -1,5 +1,7 @@
 package currency;
 
+import java.math.BigDecimal;
+
 /**
  *  Interface states important methods for a single currency pair.
  * @author Sebastian
@@ -7,6 +9,6 @@ package currency;
 public interface ExchangeData {
     public String getSenderCurrency();
     public String getReceiverCurrency();
-    public double getExchangeRate();
+    public BigDecimal getExchangeRate();
     public String getDate();
 }
